+++ src/main/java/net/more_rpg_classes/custom/spell_impacts/FrozenTicksSpellImpact.java
@@ -1,0 +1,29 @@
+package net.more_rpg_classes.custom.spell_impacts;
+
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.registry.entry.RegistryEntry;
+import net.spell_engine.api.spell.Spell;
+import net.spell_engine.api.spell.event.SpellHandlers;
+import net.spell_engine.internals.SpellHelper;
+import net.spell_power.api.SpellPower;
+
+import static net.more_rpg_classes.util.CustomMethods.stackFreezeStacks;
+
+public class FrozenTicksSpellImpact implements SpellHandlers.CustomImpact {
+
+    @Override
+    public SpellHandlers.ImpactResult onSpellImpact(
+            RegistryEntry< Spell > spell,
+            SpellPower.Result powerResult,
+            LivingEntity caster,
+            Entity target,
+            SpellHelper.ImpactContext context
+    ) {
+        if (target instanceof LivingEntity livingEntity){
+            stackFreezeStacks(livingEntity,40);
+        }
+
+        return new SpellHandlers.ImpactResult(true, false);
+    }
+}